% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook-add.R
\name{add_cover_sheet}
\alias{add_cover_sheet}
\title{Add a Cover Sheet}
\usage{
add_cover_sheet(wb, sheet_content)
}
\arguments{
\item{sheet_content}{List, required. Output from [prepare_sheet].}
}
\value{
A wbWorkbook object.
}
\description{
Add to the workbook a cover sheet. This will be the first sheet in the
workbook.
}
\examples{
\dontrun{}

}
