% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook-add.R
\name{add_tabular_sheet}
\alias{add_tabular_sheet}
\title{Add a Sheet with Tabular Content}
\usage{
add_tabular_sheet(wb, sheet_name, sheet_content)
}
\arguments{
\item{sheet_content}{List, required. Output from [prepare_sheet].}
}
\value{
A wbWorkbook object.
}
\description{
Add to the workbook a sheet with data in tabular form. This includes sheet
types 'contents', 'notes' and 'tables'. Contents and notes sheets will be
the second and third sheet in the workbook after the sheet with type 'cover'
and there can only be one of each.
}
\examples{
\dontrun{}

}
