% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-build.R
\name{append_contents}
\alias{append_contents}
\title{Add a Contents Sheet to a Blueprint}
\usage{
append_contents(
  blueprint,
  tab_name = "Contents",
  title = "Contents",
  subtitle = NULL,
  custom = NULL,
  table
)
}
\arguments{
\item{blueprint}{List, required.}

\item{tab_name}{Character vector of length 1, required.}

\item{title}{Character vector of length 1,  required. The title of the sheet.
Will be placed in cell A1 of output spreadsheet.}

\item{subtitle}{Character vector of length 1, optional, default is `NULL`. A
subtitle for the sheet. Will be placed in cell A2 of the output spreadsheet
if provided.}

\item{custom}{Named character vector, optional, default is `NULL`. Each
element will become its own row in the metadata of sheets with sheet type
'tables'. Will be placed in the order provided between the title and the
source (if provided).}

\item{table}{A data.frame, required. Tabular data. See details.}
}
\value{
A list.
}
\description{
Add a Contents Sheet to a Blueprint
}
\examples{
\dontrun{
bp <- new_blueprint() |>
  append_cover() |>
  append_contents()

bp
}

}
