% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-build.R
\name{append_cover}
\alias{append_cover}
\title{Add a Cover Sheet to a Blueprint}
\usage{
append_cover(blueprint, sheet_name = "Cover", title, subtitle = NULL, sections)
}
\arguments{
\item{blueprint}{List, required.}

\item{sheet_name}{Character vector of length 1, required.}

\item{title}{Character vector of length 1,  required. The title of the sheet.
Will be placed in cell A1 of output spreadsheet.}

\item{subtitle}{Character vector of length 1, optional, default is `NULL`. A
subtitle for the sheet. Will be placed in cell A2 of the output spreadsheet
if provided.}

\item{sections}{Named list of character vectors, required. A list-element
name becomes a section header and each character-vector element becomes row
in that section. The content will begin from cell A2 of the output
spreadsheet.}
}
\value{
A list.
}
\description{
Provide the information required to create a cover sheet.
}
\examples{
\dontrun{
bp <- new_blueprint() |>
  append_cover()

bp
}

}
