% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-build.R
\name{append_tables}
\alias{append_tables}
\title{Add a Tables Sheet to a Blueprint}
\usage{
append_tables(
  blueprint,
  sheet_name,
  title,
  subtitle = NULL,
  custom = NULL,
  source = NULL,
  tables
)
}
\arguments{
\item{blueprint}{List, required.}

\item{sheet_name}{Character vector of length 1, required.}

\item{title}{Character vector of length 1,  required. The title of the sheet.
Will be placed in cell A1 of output spreadsheet.}

\item{subtitle}{Character vector of length 1, optional, default is `NULL`. A
subtitle for the sheet. Will be placed in cell A2 of the output spreadsheet
if provided.}

\item{custom}{Named character vector, optional, default is `NULL`. Each
element will become its own row in the metadata of sheets with sheet type
'tables'. Will be placed in the order provided between the title and the
source (if provided).}

\item{source}{Character vector of length 1, optional, default is `NULL`. The
source of the data provided in the tables. Will be provided as the final
row of metadata above the tables.}

\item{tables}{A data.frame, or a list of data.frames named with table titles,
required. Tabular statistical data.}
}
\value{
A list.
}
\description{
Add a Tables Sheet to a Blueprint
}
\examples{
\dontrun{
bp <- new_blueprint() |>
  append_cover() |>
  append_contents() |>
  append_notes() |>
  append_tables() |>
  append_tables()

bp
}

}
