% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{convert_blueprint_to_workbook}
\alias{convert_blueprint_to_workbook}
\title{Convert a Blueprint to a Workbook}
\usage{
convert_blueprint_to_workbook(blueprint)
}
\arguments{
\item{blueprint}{List.}
}
\value{
An 'openxlsx' wbWorkbook- and R6-class object.
}
\description{
Accept a 'blueprint' list, possibly created with [read_blueprint], and
convert its content and structure to an 'openxlsx2' wbWorkbook-class object.
}
\examples{
# Path to an example blueprint YAML file
example_path <- system.file("extdata", "example.yaml", package = "a11ytables2")

# Read in the blueprint YAML file to a list
blueprint_list <- read_blueprint(example_path)

# Convert list to wbWorkbook-class object
workbook <- convert_blueprint_to_workbook(blueprint_list)

workbook

# openxlsx2::wb_open(workbook)  # to open temp version in spreadsheet editor

}
