% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook-build.R
\name{generate_workbook}
\alias{generate_workbook}
\title{Create a Workbook and Fill with Blueprint Information}
\usage{
generate_workbook(blueprint)
}
\value{
A wbWorkbook object.
}
\description{
Create a Workbook and Fill with Blueprint Information
}
\examples{
\dontrun{
bp <- new_blueprint() |>
  append_cover() |>
  append_contents() |>
  append_notes() |>
  append_tables() |>
  append_tables()

wb <- create_workbook(blueprint)
}

}
