% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{read_blueprint}
\alias{read_blueprint}
\title{Read a Blueprint File}
\usage{
read_blueprint(yaml_path)
}
\arguments{
\item{yaml_path}{Character. A path to a YAML file that contains a valid
blueprint for a compliant workbook. See details.}
}
\value{
A nested list. The exact content and structure will depend on the
    supplied blueprint YAML file.
}
\description{
Read to a list a 'blueprint': a YAML file that encodes the content and
structure of a workbook compliant with Analysis Function best-practice
spreadsheet guidelines.
}
\details{
The YAML file must declare certain key-value pairs to be considered valid.
}
\examples{
# Path to an example blueprint YAML file
example_path <- system.file("extdata", "example.yaml", package = "a11ytables2")

# Read in the blueprint YAML file to a list
blueprint_list <- read_blueprint(example_path)

blueprint_list

}
