% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook-build.R
\name{generate_workbook}
\alias{generate_workbook}
\title{Create a Workbook and Fill with Blueprint Information}
\usage{
generate_workbook(blueprint)
}
\arguments{
\item{blueprint}{A list, required. A blueprint for creating a workbook with
one list element for the cover, contents, notes (optional) and tables,
probably generated using [append_cover], [append_contents], [append_notes]
and [append_tables].}
}
\value{
A wbWorkbook object.
}
\description{
Create a Workbook and Fill with Blueprint Information
}
\examples{
\dontrun{
bp <- new_blueprint() |>
  append_cover() |>
  append_contents() |>
  append_notes() |>
  append_tables() |>
  append_tables()

wb <- create_workbook(blueprint)
}

}
